Move HomeAssistantTokenStorage to storage.py and remove api.py
The api.py file now only contained the HomeAssistantTokenStorage adapter
and re-exported exceptions. Since exceptions are now imported directly
from the monta package, api.py was unnecessary.

Changes:
- Create storage.py with HomeAssistantTokenStorage class
- Update imports in __init__.py to use .storage
- Update imports in config_flow.py to use .storage
- Remove api.py

This makes the module structure cleaner and more focused.
